% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Check that required columns exist in a data.table}
\usage{
check_columns(dt, required, caller = "function")
}
\arguments{
\item{dt}{data.table to check}

\item{required}{Character vector of required column names}

\item{caller}{Name of calling function for error messages}
}
\description{
Check that required columns exist in a data.table
}
\keyword{internal}
