% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_suggests}
\alias{check_suggests}
\title{Check if a suggested package is available}
\usage{
check_suggests(pkg, reason = NULL)
}
\arguments{
\item{pkg}{Package name}

\item{reason}{Why the package is needed}
}
\description{
Check if a suggested package is available
}
\keyword{internal}
