% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{cluster_professions}
\alias{cluster_professions}
\title{Hierarchical clustering of professions}
\usage{
cluster_professions(distance, method = "complete")
}
\arguments{
\item{distance}{A \code{dist} object, typically from
\code{\link[=compute_profession_distance]{compute_profession_distance()}}.}

\item{method}{Character, agglomeration method passed to
\code{\link[stats:hclust]{stats::hclust()}}. Default: \code{"complete"}. Other options:
\code{"ward.D2"}, \code{"average"}, \code{"single"}, etc.}
}
\value{
An \code{hclust} object.
}
\description{
Wrapper around \code{\link[stats:hclust]{stats::hclust()}} for clustering a profession distance
matrix. Returns the \code{hclust} object which can be plotted as a
dendrogram or cut into groups with \code{\link[stats:cutree]{stats::cutree()}}.
}
