% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isco_gruppi}
\alias{isco_gruppi}
\title{ESCO ISCO group classification}
\format{
A data.table with columns:
\describe{
\item{code}{ESCO occupation code}
\item{preferredLabel}{Profession name (Italian)}
\item{esco_v0101_hier_label_1}{Hierarchy level 1}
\item{esco_v0101_hier_label_2}{Hierarchy level 2}
\item{esco_v0101_hier_label_3}{Hierarchy level 3}
}
}
\source{
ESCO classification v1.1
}
\usage{
isco_gruppi
}
\description{
Mapping table from ESCO level 4 occupation codes to profession names
and hierarchy labels.
}
\keyword{datasets}
