% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skillviz-package.R
\docType{package}
\name{skillviz-package}
\alias{skillviz}
\alias{skillviz-package}
\title{skillviz: Skill Co-Occurrence Analysis and Visualization from Online Job Advertisements}
\description{
Analyzes skill requirements from Online Job Advertising (OJA) data. Provides tools for computing skill specialization indices (Balassa/RCA), co-occurrence networks, source stability filtering, salary analysis, professional distance clustering, temporal skill ranking, emerging skill indicators, profile change of professions and share forecasting. Built on ESCO and CP2021 (Italian professional classifications) taxonomies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gmontaletti.github.io/skillviz/}
  \item \url{https://github.com/gmontaletti/skillviz}
  \item Report bugs at \url{https://github.com/gmontaletti/skillviz/issues}
}

}
\author{
\strong{Maintainer}: Giampaolo Montaletti \email{giampaolo.montaletti@gmail.com} (\href{https://orcid.org/0009-0002-5327-1122}{ORCID})

}
\keyword{internal}
